requirements_Version2.txt
tkinter
fpdf2
pyspellchecker
python-docx